Introduce concept of Tray
This is where the pieces will spawn, overlapped.
The player then moves them to the board.

The board and tray are fitted next to / on top of each other, and shrunk to both fit fully on the screen.

I am not sure if I like this approach yet.
But I'm keeping it like this for the time being.